responses packages

+++ api/src/repositories/users.go
@@ -9,14 +9,13 @@ type users struct {
 	db *sql.DB
 }
 
-//NewUsersRepositories creates a user repository
+// NewUsersRepositories creates a user repository
 func NewUsersRepositories(db *sql.DB) *users {
 	return &users{db}
 }
 
-
-//Create inserts a user into the database
-func (u users) Create(user models.User) (uint32, error) {
+// Create inserts a user into the database
+func (u users) Create(user models.User) (uint64, error) {
 	statment, err := u.db.Prepare("insert into users (name, nick, email, password) values(?, ?, ?, ?)")
 	if err != nil {
 		return 0, err
@@ -28,11 +27,10 @@ func (u users) Create(user models.User) (uint32, error) {
 		return 0, err
 	}
 
-
 	lastEnteredID, err := result.LastInsertId()
 	if err != nil {
 		return 0, err
 	}
 
-	return uint32(lastEnteredID), nil
+	return uint64(lastEnteredID), nil
 }
